Posts now have a user associated with them.

+++ src/Blog.hs
@@ -25,6 +25,7 @@ import           Network.HTTP.Types.Status
 import qualified Database.Redis as Redis
 import           Database.PostgreSQL.Simple as PG
 import           Database.PostgreSQL.Simple.Migration as PG.Migration
+import           Database.PostgreSQL.Simple.ToField as PG.ToField
 
 import           Text.Blaze.Html5 as H hiding (style, param, map)
 import           Text.Blaze.Html5.Attributes as A
@@ -34,19 +35,15 @@ import           Prelude as P hiding (head, id, div)
 import           Magic -- for mimetypes
 
 -- TODO:
--- 1. Finish pages (procrastinateable)
---    a. not found
---    b. unauthorized
--- 2. Page numbers at bottom (would require extra request)
--- 4. drafts
--- 5. blog post author
--- 6. Site footer (copyright etc)
--- 7. Finish mobilization: Fixing tags in editor
+-- Page numbers at bottom (would require extra db hit)
+-- drafts
+-- Site footer (copyright etc)
 
 -- Nitpick todo:
--- 1. fix timezone
--- 2. Rewrite using a state monad?
--- 4. cache minification results
+-- fix timezone
+-- Rewrite using a state monad?
+-- rejigger DB to use queries like SELECT b.identifier, b.title, u FROM blogposts b, (SELECT * FROM users WHERE id=1) u; instead of storing duplicate users in the DB
+-- SELECT b.identifier, b.title, u FROM blogposts b, users u WHERE u.id=(b.author).id
 
 -- Miscellania:
 -- 1. 'Top 5' tags map in side bar?
@@ -68,9 +65,10 @@ main = do
     liftIO $ execute_ pg "SET client_min_messages=NOTICE;"
     liftIO $ withTransaction pg $ runMigration $ MigrationContext (MigrationFile "blog.sql" "./migrations/blog.sql") True pg
     liftIO $ withTransaction pg $ runMigration $ MigrationContext (MigrationFile "array_funcs.sql" "./migrations/array_funcs.sql") True pg
-    liftIO $ putStrLn "--| clearing cached data from Redis"
-    liftIO $ Redis.runRedis redis $ Redis.flushall
-    liftIO $ putStrLn "--| done |--"
+    liftIO $ withTransaction pg $ runMigration $ MigrationContext (MigrationFile "authorship.sql" "./migrations/authorship.sql") True pg
+    -- liftIO $ putStrLn "--| clearing cached data from Redis"
+    -- liftIO $ Redis.runRedis redis $ Redis.flushall
+    liftIO $ putStrLn "--| blog started"
   
     -- shortcuts for auth
     let protected = checkAuth redis
@@ -105,7 +103,7 @@ main = do
     get "/posts/:id" $ do
       identifier_ <- param "id"
       maybeUser <- authenticatedUser
-      res <- liftIO $ listToMaybe <$> query pg "SELECT * FROM blogposts WHERE identifier=? LIMIT 1" [identifier_ :: Integer]
+      res <- liftIO $ getBlogPost pg identifier_
       case res of
         Nothing -> next
         Just post_ -> do
@@ -123,7 +121,6 @@ main = do
       Scotty.html $ R.renderHtml $ docTypeHtml $ do
         renderHead [] [] (appendedBlogTitle $ tag)
         renderBody (Just $ T.append "Posts tagged '" $ T.append tag "'") Nothing maybeUser $ do
-          -- render (posts :: [BlogPost]) maybeUser
           render (take postsPerPage posts) maybeUser
           renderPageControls mPageNum (length posts > postsPerPage)
   
@@ -131,7 +128,7 @@ main = do
     get "/posts/:id/edit" $ do
       protected
       identifier_ <- param "id"
-      res <- liftIO $ listToMaybe <$> query pg "SELECT * FROM blogposts WHERE identifier=? LIMIT 1" [identifier_ :: Integer]
+      res <- liftIO $ getBlogPost pg identifier_
       case res of
         Nothing -> next
         Just post_ -> do
@@ -167,13 +164,13 @@ main = do
       case res of
         Nothing     -> redirect "/login?error_message=Username%20not%20found%2E"
         (Just user) -> do
-          if Helpers.checkPassword (Types.passwordHash user) pPassword
+          if Helpers.checkPassword (fromJust $ Types.passwordHash user) pPassword
             then do
               token <- liftIO $ Auth.saveObject redis user
               setAccessToken (T.decodeUtf8 token)
               redirect "/"
             else redirect "/login?error_message=Invalid%20password%2E"
-        
+    
     matchAny "/unauthorized" $ do
       status $ Status 401 "Unauthorized"
       Scotty.html $ R.renderHtml $ docTypeHtml $ do
@@ -184,30 +181,35 @@ main = do
     -- deletes a BlogPost from the database
     -- returns JSON
     delete "/posts/:id" $ do
-      protected
+      authUser <- protected
       identifier_ <- param "id"
-      res <- liftIO $ listToMaybe <$> query pg "DELETE FROM blogposts WHERE identifier=? RETURNING *" [identifier_ :: Integer]
-      case res of
-        Nothing -> emptyResponse
-        Just bp -> Scotty.json (bp :: BlogPost)
-  
+      res <- liftIO $ listToMaybe <$> query pg "DELETE FROM blogposts WHERE identifier=? AND (author).id=? RETURNING *" (identifier_ :: Integer, Types.uid $ fromJust authUser)
+      case (res :: Maybe BlogPost) of
+        Nothing -> do
+          status $ Status 404 "blog post not found."
+        Just bp -> do
+          Scotty.text "ok"  
+
     -- creates/updates a BlogPost in the database
     -- returns JSON
     post "/posts" $ do
-      protected
+      authUser <- protected
       ps <- params
-      maybeBlogPost <- liftIO $ (upsertBlogPost pg
-                                               (((read . TL.unpack) <$> lookup "id" ps) :: Maybe Integer)
-                                               (TL.toStrict <$> lookup "title" ps)
-                                               (TL.toStrict <$> lookup "body" ps)
-                                               (((splitList ',') . TL.unpack) <$> (lookup "tags" ps))
-                                               (((splitList ',') . TL.unpack) <$> (lookup "deleted_tags" ps))
-                                              )
-      case maybeBlogPost of
-        Nothing -> emptyResponse
-        Just bp -> do
-          addHeader "Location" $ TL.pack $ "/posts/" ++ (show $ Types.identifier bp)
-          Scotty.json (bp :: BlogPost)
+      maybeBPIdentifier <- liftIO $ (upsertBlogPost pg
+                                                    (fromJust authUser)
+                                                    (((read . TL.unpack) <$> lookup "id" ps) :: Maybe Integer)
+                                                    (TL.toStrict <$> lookup "title" ps)
+                                                    (TL.toStrict <$> lookup "body" ps)
+                                                    (((splitList ',') . TL.unpack) <$> (lookup "tags" ps))
+                                                    (((splitList ',') . TL.unpack) <$> (lookup "deleted_tags" ps))
+                                                   )
+      case maybeBPIdentifier of
+        Nothing -> do
+          status $ Status 400 "Missing required parameters"
+          emptyResponse
+        Just identifier_ -> do
+          addHeader "Location" $ TL.pack $ "/posts/" ++ (show identifier_)
+          emptyResponse
   
     -- returns minified JS
     get "/assets/js/:filename" $ do
@@ -283,17 +285,17 @@ renderMdEditor :: Maybe BlogPost -> Html
 renderMdEditor Nothing = do
   div ! A.id "preview" $ ""
   textarea ! A.id "editor" $ ""
-renderMdEditor (Just (BlogPost identifier_ _ body_ _ _)) = do
+renderMdEditor (Just (BlogPost identifier_ _ body_ _ _ _ _)) = do
   div ! A.id "preview" $ ""
   textarea ! A.id "editor" ! customAttribute "post-id" (stringValue $ show identifier_) $ toHtml body_
       
 renderTitleField :: Maybe BlogPost -> Html
-renderTitleField (Just (BlogPost _ title_ _ _ _)) = input ! type_ "text" ! id "title-field" ! placeholder "Post title" ! value (textValue title_)
+renderTitleField (Just (BlogPost _ title_ _ _ _ _ _)) = input ! type_ "text" ! id "title-field" ! placeholder "Post title" ! value (textValue title_)
 renderTitleField Nothing = input ! type_ "text" ! id "title-field" ! placeholder "Post title"
 
 renderTagEditor :: Maybe BlogPost -> Html
 renderTagEditor Nothing = textarea ! A.id "tags" ! class_ "wordlist" $ do ""
-renderTagEditor (Just (BlogPost _ _ _ _ tags_)) = textarea ! A.id "tags" ! class_ "wordlist" $ do toHtml $ List.intercalate ", " tags_
+renderTagEditor (Just (BlogPost _ _ _ _ tags_ _ _)) = textarea ! A.id "tags" ! class_ "wordlist" $ do toHtml $ List.intercalate ", " tags_
   
 renderPostEditor :: Maybe BlogPost -> Html
 renderPostEditor maybeBlogPost = do
@@ -373,29 +375,35 @@ cachedBody redis key valueFunc = do
 -------------------------------------------------------------------------------
 --- | Database
 
-upsertBlogPost :: PG.Connection -> Maybe Integer -> Maybe T.Text -> Maybe T.Text -> Maybe [String] -> Maybe [String] -> IO (Maybe BlogPost)
-upsertBlogPost pg (Just identifier_) Nothing       Nothing      Nothing      Nothing             = listToMaybe <$> query pg "SELECT * FROM blogposts WHERE identifier=? LIMIT 1" [identifier_]
-upsertBlogPost pg (Just identifier_) (Just title_) Nothing      Nothing      Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET title=? WHERE identifier=? RETURNING *" (title_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) (Just body_) Nothing      Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET title=?, bodyText=? WHERE identifier=? RETURNING *" (title_, body_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) (Just body_) (Just tags_) Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET title=?, bodyText=?, tags=uniq_cat(tags,?)  WHERE identifier=? RETURNING *" (title_, body_, tags_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) (Just body_) (Just tags_) (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET title=?, bodyText=?, tags=array_diff(uniq_cat(tags, ?),?) WHERE identifier = ? RETURNING *" (title_, body_, tags_, deletedTags_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       (Just body_) Nothing      Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET bodyText=? WHERE identifier=? RETURNING *" (body_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       Nothing      (Just tags_) Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET tags=uniq_cat(tags,?) WHERE identifier=? RETURNING *" (tags_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       (Just body_) (Just tags_) Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET bodyText=?, tags=uniq_cat(tags,?) WHERE identifier=? RETURNING *" (body_, tags_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) Nothing      (Just tags_) Nothing             = listToMaybe <$> query pg "UPDATE blogposts SET title=?, tags=uniq_cat(tags,?) WHERE identifier=? RETURNING *" (title_, tags_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       Nothing      Nothing      (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET tags=array_diff(tags,?) WHERE identifier=? RETURNING *" (deletedTags_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       Nothing      (Just tags_) (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET tags=uniq_cat(array_diff(tags,?),?) WHERE identifier=? RETURNING *" (deletedTags_, tags_, identifier_)
-upsertBlogPost pg (Just identifier_) Nothing       (Just body_) Nothing      (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET bodyText=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING *" (body_, deletedTags_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) Nothing      Nothing      (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET title=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING *" (title_, deletedTags_, identifier_)
-upsertBlogPost pg (Just identifier_) (Just title_) (Just body_) Nothing      (Just deletedTags_) = listToMaybe <$> query pg "UPDATE blogposts SET title=?, bodyText=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING *" (title_, body_, deletedTags_, identifier_)
-upsertBlogPost pg Nothing            (Just title_) (Just body_) Nothing      _                   = listToMaybe <$> query pg "INSERT INTO blogposts (title, bodyText) VALUES (?, ?) RETURNING *" (title_, body_)
-upsertBlogPost pg Nothing            (Just title_) (Just body_) (Just tags_) _                   = listToMaybe <$> query pg "INSERT INTO blogposts (title, bodyText, tags) VALUES (?, ?, ?) RETURNING *" (title_, body_, tags_)
-upsertBlogPost _  _                 _            _            _           _                      = return Nothing
+upsertBlogPost :: PG.Connection -> User -> Maybe Integer -> Maybe T.Text -> Maybe T.Text -> Maybe [String] -> Maybe [String] -> IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       Nothing      Nothing      Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "SELECT identifier FROM blogposts WHERE identifier=? LIMIT 1" [identifier_]) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) Nothing      Nothing      Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=? WHERE identifier=? RETURNING identifier" (title_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) (Just body_) Nothing      Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, bodyText=? WHERE identifier=? RETURNING identifier" (title_, body_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) (Just body_) (Just tags_) Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, bodyText=?, tags=uniq_cat(tags,?)  WHERE identifier=? RETURNING identifier" (title_, body_, tags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) (Just body_) (Just tags_) (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, bodyText=?, tags=array_diff(uniq_cat(tags, ?),?) WHERE identifier = ? RETURNING identifier" (title_, body_, tags_, deletedTags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       (Just body_) Nothing      Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET bodyText=? WHERE identifier=? RETURNING identifier" (body_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       Nothing      (Just tags_) Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET tags=uniq_cat(tags,?) WHERE identifier=? RETURNING identifier" (tags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       (Just body_) (Just tags_) Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET bodyText=?, tags=uniq_cat(tags,?) WHERE identifier=? RETURNING identifier" (body_, tags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) Nothing      (Just tags_) Nothing             = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, tags=uniq_cat(tags,?) WHERE identifier=? RETURNING identifier" (title_, tags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       Nothing      Nothing      (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET tags=array_diff(tags,?) WHERE identifier=? RETURNING identifier" (deletedTags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       Nothing      (Just tags_) (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET tags=uniq_cat(array_diff(tags,?),?) WHERE identifier=? RETURNING identifier" (deletedTags_, tags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) Nothing       (Just body_) Nothing      (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET bodyText=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING identifier" (body_, deletedTags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) Nothing      Nothing      (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING identifier" (title_, deletedTags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user (Just identifier_) (Just title_) (Just body_) Nothing      (Just deletedTags_) = (listToMaybe <$> map fromOnly <$> ((query pg "UPDATE blogposts SET title=?, bodyText=?, tags=array_diff(tags,?) WHERE identifier=? RETURNING identifier" (title_, body_, deletedTags_, identifier_)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user Nothing            (Just title_) (Just body_) Nothing      _                   = (listToMaybe <$> map fromOnly <$> ((query pg "INSERT INTO blogposts (title, bodyText, author_id) VALUES (?, ?, ?) RETURNING identifier" (title_, body_, Types.uid user)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost pg user Nothing            (Just title_) (Just body_) (Just tags_) _                   = (listToMaybe <$> map fromOnly <$> ((query pg "INSERT INTO blogposts (title, bodyText, tags, author_id) VALUES (?, ?, ?, ?) RETURNING identifier" (title_, body_, tags_, Types.uid user)) :: IO [Only Integer])) :: IO (Maybe Integer)
+upsertBlogPost _  _    _                 _            _            _           _                      = return Nothing
 
 getBlogPostsByTag :: PG.Connection -> T.Text -> Maybe Integer -> IO [BlogPost]
 getBlogPostsByTag pg tag Nothing        = getBlogPostsByTag pg tag (Just 1)
-getBlogPostsByTag pg tag (Just pageNum) = query pg "SELECT * FROM blogposts WHERE ?=any(tags) ORDER BY identifier DESC OFFSET ? LIMIT ?" (tag,(pageNum-1)*(fromIntegral postsPerPage),postsPerPage+1)
+getBlogPostsByTag pg tag (Just pageNum) = query pg "SELECT b.*,u FROM blogposts b, users u WHERE u.id=b.author_id WHERE ?=any(tags) ORDER BY identifier DESC OFFSET ? LIMIT ?" (tag,(pageNum-1)*(fromIntegral postsPerPage),postsPerPage+1)
 
 getBlogPosts :: PG.Connection -> Maybe Integer -> IO [BlogPost]
-getBlogPosts pg (Just pageNum) = query pg "SELECT * FROM blogposts ORDER BY identifier DESC OFFSET ? LIMIT ?" ((pageNum-1)*(fromIntegral postsPerPage), postsPerPage+1)
+getBlogPosts pg (Just pageNum) = query pg "SELECT b.*,u FROM blogposts b, users u WHERE u.id=b.author_id ORDER BY identifier DESC OFFSET ? LIMIT ?" ((pageNum-1)*(fromIntegral postsPerPage), postsPerPage+1)
 getBlogPosts pg Nothing        = getBlogPosts pg (Just 1)
+
+getBlogPost :: PG.Connection -> Integer -> IO (Maybe BlogPost)
+getBlogPost pg identifier_ = listToMaybe <$> query pg "SELECT b.*,u FROM blogposts b, users u WHERE u.id=b.author_id AND identifier=? LIMIT 1" [identifier_ :: Integer]
+
+-- getPageCount :: PG.Connection -> IO (Maybe Integer)
+-- getPageCount pg = listToMaybe <$> query_ pg "SELECT ceil(COUNT(*)::real/10::real) as page_count FROM blogposts"
